requirements.txt
Web3
python-dotenv